Implemented the changes that were part of 20275 in the server.
By doing this I noticed a bug that was introduced as part of turning
rabbit_channel into a gen_server2.

Previously when the broker sent the connection.close, this was being handled
by a handle_info callback in the amqp_connection module.

To compound this, there was a stupid cast handler that swallowed any casts,
thus preventing any diagnostics from bubbling up.

--HG--
branch : bug20321