Introduced a closing state when the broker initiates the shutdown. This allows the socket reader process to wait for the broker to close the socket (or timeout) before stopping amqp_connection process. Squirrel the close reason away in the state of the connection process. The resulting EXIT message from the reader process is trapped and then the close reason that was squirreled away can is used as the EXIT reason of the amqp_connection process.
Because the amqp_channel process links to the connection process, it will
receive the same EXIT reason and return it to whatever application
process is using the amqp_channel process.

The network driver sees some minor defintion cleanups as well as removing
the link between the amqp_channel_process and the rabbit_framing_channel
process -
this is because the amqp_channel is not able to sensibly handle any exit
the framing channel will send.

Made the teardown routine assert that the channel and connection
processes are in actual fact dead after having closed the connection for every test.

The connection process does not shutdown synchronously.

Also put some repetitions into the tests that were experiencing race conditions.

--HG--
branch : bug21522